backend/requirements.txt
telethon
python-dotenv
